Small tweaks for WebDavClientService and tests
